# gateway-service/api/openapi.yaml
openapi: 3.0.3
info:
  title: NEXCOM Exchange Gateway API
  description: |
    REST API for the NEXCOM Next-Generation Commodity Exchange platform.
    Provides endpoints for trading, market data, portfolio management,
    clearing, surveillance, and platform administration.
  version: 1.0.0
  contact:
    name: NEXCOM Exchange
    url: https://nexcom.exchange
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.nexcom.exchange/api/v1
    description: Production
...
paths:
...
  /analytics/dashboard:
    get:
      summary: Analytics dashboard data
      operationId: analyticsDashboard
      tags: [Analytics]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard metrics
  /analytics/pnl:
    get:
      summary: P&L report
      operationId: pnlReport
      tags: [Analytics]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: P&L data
  /analytics/geospatial/{commodity}:
    get:
      summary: Geospatial data for a commodity
      operationId: geospatialData
      tags: [Analytics]
      security:
        - bearerAuth: []
      parameters:
        - name: commodity
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Geospatial analytics
  /analytics/ai-insights:
    get:
      summary: AI/ML market insights
      operationId: aiInsights
      tags: [Analytics]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: AI-generated insights
  /analytics/forecast/{symbol}:
    get:
      summary: Price forecast for a symbol
      operationId: priceForecast
      tags: [Analytics]
      security:
        - bearerAuth: []
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Price forecast with confidence intervals
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak OIDC JWT token
...